Add files via upload
This is a simple Note Taking Application built using Flask, a Python web framework. The app provides a simple and intuitive interface that allows users to add new notes, update the existing ones or even delete some of them as per user's choice. This work on Notefy app is the project/task no.4 of my internship period at Innomatics Researcg Labs as a Data Science Intern in Feb 2024.